Updates in the Arabic locale to match those in the English version
After the release of OJS ver. 2.4.7, some locale files were updated in
the English Language by adding several keys here and there, so the
Arabic version should be updated as well.
I made a brief revision to the styles used in the system, and found that
some adjustments were necessary to have a better experience to the
Arabic user. I added the (div) tag there because the right-to-left
orientation in the (body) tag wasn't enough the satisfy the required
page content view.