pkp/pkp-lib#1443 Restructure site/context registration
This commit is part of restructuring that removes the
userRegisterSite.tpl and passes all frontend
registration through the same template. It also
distinguishes the role signup process depending on
whether the user is registering with a journal or
with the overall site. If it is with the overall site,
all roles are selected manually. If a context exists,
the user is opted into the Reader and Author roles.