pkp/pkp-lib#1468 Add {load_header} smarty function.
This commit adds the {load_header} smarty function which provides a
common api for adding assets to the HTML header of a page which may
not be covered by the script and style apis.

In addition, a new technique for loading page- and op-sensitive
contexts was added. Now, when loading the 'frontend' context on
a single issue view, additional dynamic contexts will be checked
for registered assets according to the requested Page and Op. In
the single issue view example, it would check for assets on the
'frontend-issue' and 'frontend-issue-view' contexts. This allows
scripts, styles and header items to be loaded in a more
granular way, which is also exposed to third-party plugins which
may want to add or modify the pages in which items are loaded.