Check is_array again
If $this->_data[$key] isn't an array, then you might get something like:

stderr: PHP Warning:  Cannot use a scalar value as an array in ojs3/lib/pkp/classes/core/DataObject.inc.php on line 134, referer: http://ojs3.localhost//admin/settings

This causes the settings page, in the first tab (settings), to be blank (and the "loading" icon to stay forever loading). This is related with custom locales (we are using es_AR, which seems to be very incomplete at the time).