pkp/pkp-lib#1740 Prevent file browser from opening twice in Windows browsers.
This commit removes code which passes clicks on the file upload
button to the plupload control. This code was added to improve
keyboard accessibility, but in Windows browsers it caused the
file browser to be opened twice.

Removing this code presents accessibility challenges, which were
addressed by providing visual feedback which mocked a focus
effect on the browse button when focus is placed on plupload's
control. This does not address the accessibility issue whereby
the plupload control has no meaningful text label. Screen readers
may suffer.

In addition, disabling clicks on the visual button could cause
some compatibility runtimes to break. plupload includes several
different compatibility runtimes to ensure as wide a browser
support as possible. But we're not able to test all of these
comprehensively. We may find the downside to this change is
bigger than the upside.