pkp/pkp-lib#2163 Refactor user merging to avoid partial handler refresh
Previously, a user tab handler performed a partial tab refresh
to refresh the user grid when a user merge occurred. This
causes problems for the handler event deregistration process
I'm working on implementing. This commit refactors away from
that approach in favor of a modal approach. It also fixed
pkp/pkp-lib#2171 in the process.