pkp/pk-lib#2163 Use an offset instead of paging for submissions lists.
This commit fixes an issue with the traditional paging technique for loading
sets of submissions into the new submissions list component. When a submission
is deleted, the Load More feature on the component needs to start at an
offset which doesn't match the exact page set.

To address this, DBResultRange has been extended with an offset option and
DAO::retrieveRange() will use SelectLimit instead of PageExecute when retrieving
the result set. However, to get the total count of items, we have to execute
a secondary paged query.