pkp/pkp-lib#2163 Implement split between pkp-lib and apps for api and services.
This commit splits the code for the API and the service classes to support
the division of labor between the pkp-lib and omp/ojs. It creates new app-
specific classes for the API handler, submission services, services
container, and query builder, which extend base classes prefixed with PKP.

It also restructures the Backend API handler to use contextual endpoints.
Instead of /backend/submissions, the submissions endpoints can now be found
at /_submissions. This will prevent having to put all of our private
endpoints in a single class, and should make it easier to extend them for
each app.

Finally, this commit introduces filtering in ListPanel components. This is
used for the catalog management list in OMP, but should be reasonably
re-usable for other ListPanels. The changes introduced to the ListPanel
are still a work-in-progress.