pkp/pkp-lib#2894 Implement searching by interest in reviewer selection.
This commit extends the searchPhrase to checking user interests. In the
process, it also exposed a bug with the reviewer selection sums on
completed and active reviews. To address this, the sums needed to be
placed into separate subqueries to prevent multiple lines from the
search phrase left joins from inflating the sum counts.