pkp/pkp-lib#2894 Add reviewer gossip field.
This commit restores the ability to add editorial notes about a reviewer.
These notes can be viewed/added by admins, journal managers, editors and
subeditors. The field is available in the edit user modal as well as from
the reviewer assignment grid. No one is allowed to view their own gossip
field.

It is not possible to match reviewer search phrases with the gossip field.
Making sure that matches wouldn't be possible for one's own gossip field
could require some complex SQL in UserListQueryBuilder.

I decided not to allow anyone to see their own gossip field as a stop-gap
in case a reviewer is gossiped about, then later is promoted to an
editorial position. This way, there won't be old comments mistakenly
exposed.