Added the possibility for plugins to provide their own help files.
The plugin base class was extended to register a plugins help index file
to the Help::Plugins hook.

The PKPTemplateManager smartyHelp method was extended to support the
parameter plugin, to create a plugin help link, defaulting to index.md
as filename.

The HelpHandler now handles plugin help chapters as well.