make HookRegistry ignore non-callable callbacks
If display_errors=on, a broken plugin (such as OpenAIREPlugin
in my case) that registers something that cannot be called,
will provoke a message that breaks OJS's AJAX pages (because
many JSON responses will become ill-formed).
This patch ignores hook callbacks that are !is_callable
and so makes OJS robust against such cases.